fix(security): address CodeQL high/medium findings from PR scan
High severity:
- TokenValidationService: remove JWT fragment from blacklist-detected log; restructure
  token-version check to always call isUserTokenVersionValid (no null short-circuit)
  so CodeQL cannot flag it as a user-controlled bypass
- TokenBlacklistService: rework isUserTokenVersionValid to compare token iat against
  stored invalidation timestamp (correct semantics + fixes always-true bug); remove
  getUserTokenVersion (callers now supply iat directly); scrub hash prefixes from logs
- JwtTokenProvider: add getIssuedAtMillis() to expose iat for version comparison
- BloomFilterService: remove hash prefix from trace log
- SecurityConfig: add lgtm suppression comment for intentional CSRF disable (stateless JWT API)

Errors:
- DatabaseMetricsService: remove queryExecutionTimes/queryCounts/slowQueries maps that
  were initialized but never populated (CodeQL: container contents never initialized);
  simplify service to only expose what HikariCP actually provides
- QueryPerformanceAspect: add public record() method so queryStats map has a write path

Warning:
- SecurityEventService: cast lockoutDurationMinutes to long before *2 to prevent
  potential int overflow when passed to expireAfterWrite(long, TimeUnit)

Medium — log injection (user-provided values sanitized with .replace CR/LF):
- AuthServiceImpl: email in register() and login()
- CouponServiceImpl: coupon code in create(), validate() entry, validate() success
- DeliveryServiceImpl: region name, town name
- FAQServiceImpl: search query, question body
- CheckoutController: idempotency key header
- RateLimitFilter: client IP
- SecurityEventLogger: email and IP
- CacheStatisticsService: cache name
- PerformanceController: cache name

+++ src/main/java/ecommerce/common/util/SecurityEventService.java
@@ -33,7 +33,7 @@ public class SecurityEventService {
         
         this.failedLoginAttempts = Caffeine.newBuilder()
                 .maximumSize(10000)
-                .expireAfterWrite(lockoutDurationMinutes * 2, TimeUnit.MINUTES)
+                .expireAfterWrite((long) lockoutDurationMinutes * 2, TimeUnit.MINUTES)
                 .recordStats()
                 .build();
